Deprecate Uniquify and move its useful utilities
